+++ login-utils-mgr/src/main/java/org/loginutils/mgr/controller/user/AddUserRequest.java
@@ -34,4 +34,6 @@ class AddUserRequest {
     private String email;
 
     private String phone;
+    
+    private String language;
 }