+++ login-utils-mgr/src/main/java/org/loginutils/mgr/service/LoginService.java
@@ -90,6 +90,7 @@ public class LoginService {
                 .userId(user.getUserId())
                 .username(user.getUsername())
                 .status(user.getStatus())
+                .language(user.getLanguage())
                 .roles(roles)
                 .build();